# enhanced self-documentation for inspect functions