This PR improves the readability of code for inspecting functions.
When I tried to understand inspecting of functions, looking into source code, I could not manage to get my head around implementation details. Maybe this helps future devs a bit.

I did not yet investigate perfornmance; considerations are welcome.


All private functions tell what they do by the name but
```elixir
exported?(mod)
``` 
I highly appreciate naming suggestions.

Same tests fail before and after.