@josevalim 2023-08-11
Hi @kkalb, thank you for the PR. However, we tend to not accept PRs that are refactoring code via renaming or extracting non-duplicated functions, because often it boils down to style and preferences. For example, it is unclear that `inspect_atom` is a better name than `inspected_as_atom`. Also, further changes, such as encapsulating `to_str` makes the code less efficient and so on. For those reasons, we won't merge it. :)