@kkalb 2023-08-11
> Hi @kkalb, thank you for the PR. However, we tend to not accept PRs that are refactoring code via renaming or extracting non-duplicated functions, because often it boils down to style and preferences. For example, it is unclear that `inspect_atom` is a better name than `inspected_as_atom`. Also, further changes, such as encapsulating `to_str` makes the code less efficient and so on. For those reasons, we won't merge it. :)

Understood!

So only bugfixes, new features and new or improved tests are wanted from non-core contributors?

With respecting existing code style of the refactored code and not introducing lesser efficient code, is there a way on improving the code base actively?

I am asking because refactoring PRs are the most welcome for our companies projects. Switching between this contexts is not a no-brainer imho.

I assume performance > readability/docu is key on the core of the language.
Even if I can not confirm that it has an measureable performance impact, I know that it is there.
Or I just measure it wrong?

**Without to_str() function**

```bash
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1565689    0.52
Inspect.Function.__impl__/1 1000000  6.71  105114    0.11
Inspect.Function.uniq/1     1000000 20.20  316328    0.32
Inspect.Function.inspect/2  1000000 73.08 1144247    1.14

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1578157    0.53
Inspect.Function.__impl__/1 1000000  6.70  105788    0.11
Inspect.Function.uniq/1     1000000 20.33  320794    0.32
Inspect.Function.inspect/2  1000000 72.97 1151575    1.15

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1552247    0.52
Inspect.Function.__impl__/1 1000000  6.81  105638    0.11
Inspect.Function.uniq/1     1000000 20.26  314471    0.31
Inspect.Function.inspect/2  1000000 72.94 1132138    1.13

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1549106    0.52
Inspect.Function.__impl__/1 1000000  6.71  103945    0.10
Inspect.Function.uniq/1     1000000 21.03  325788    0.33
Inspect.Function.inspect/2  1000000 72.26 1119373    1.12

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1578853    0.53
Inspect.Function.__impl__/1 1000000  6.61  104431    0.10
Inspect.Function.uniq/1     1000000 20.53  324201    0.32
Inspect.Function.inspect/2  1000000 72.85 1150221    1.15

```

**With to_str() function**

```bash

Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1519723    0.51
Inspect.Function.__impl__/1 1000000  6.80  103309    0.10
Inspect.Function.uniq/1     1000000 20.25  307721    0.31
Inspect.Function.inspect/2  1000000 72.95 1108693    1.11

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1590583    0.53
Inspect.Function.__impl__/1 1000000  6.82  108413    0.11
Inspect.Function.uniq/1     1000000 20.35  323761    0.32
Inspect.Function.inspect/2  1000000 72.83 1158409    1.16

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1589522    0.53
Inspect.Function.__impl__/1 1000000  6.78  107796    0.11
Inspect.Function.uniq/1     1000000 20.16  320439    0.32
Inspect.Function.inspect/2  1000000 73.06 1161287    1.16

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1587273    0.53
Inspect.Function.__impl__/1 1000000  6.71  106475    0.11
Inspect.Function.uniq/1     1000000 20.22  320903    0.32
Inspect.Function.inspect/2  1000000 73.07 1159895    1.16

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1555989    0.52
Inspect.Function.__impl__/1 1000000  6.91  107479    0.11
Inspect.Function.uniq/1     1000000 20.25  315073    0.32
Inspect.Function.inspect/2  1000000 72.84 1133437    1.13

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1539055    0.51
Inspect.Function.__impl__/1 1000000  6.76  104109    0.10
Inspect.Function.uniq/1     1000000 20.62  317428    0.32
Inspect.Function.inspect/2  1000000 72.61 1117518    1.12

Profile done over 3 matching functions
VirtualBox:~/elixir$ mix profile.eprof -e "Enum.each(0..999999, fn _ ->  inspect((fn arg -> arg end)) end)" --matching Inspect.Function
Warmup...


Profile results of #PID<0.109.0>
#                             CALLS     %    TIME µS/CALL
Total                       3000000 100.0 1550124    0.52
Inspect.Function.__impl__/1 1000000  7.16  110945    0.11
Inspect.Function.uniq/1     1000000 20.31  314772    0.31
Inspect.Function.inspect/2  1000000 72.54 1124407    1.12

Profile done over 3 matching functions
```